refine conditional compilation for mem control on different OSs
